Migrate BoxClassTransformer to its own class and implement missing methods
